requirements.txt
beautifulsoup4~=4.12.3
python-allrecipes~=0.3.1